requirements.txt
astroid==3.3.4
build==1.2.2
click==8.1.7
coverage==7.6.1
dill==0.3.9
greenlet==3.1.1
iniconfig==2.0.0
isort==5.13.2
mccabe==0.7.0
mysql-connector-python==9.0.0
mysql-connector-python-rf==2.2.2
packaging==24.1
pip-tools==7.4.1
platformdirs==4.3.6
pluggy==1.5.0
pylint==3.3.1
PyMySQL==1.1.1
pyproject_hooks==1.2.0
pytest==8.3.3
pytest-cov==5.0.0
python-dotenv==1.0.1
setuptools==75.1.0
SQLAlchemy==2.0.35
SQLAlchemy-Utils==0.41.2
tomlkit==0.13.2
typing_extensions==4.12.2
wheel==0.44.0
